+++ webui/src/misc/eventHandler.js
@@ -1,0 +1,6 @@
+export const subscribeOnTargetEvent = (target, event, handler) => {
+  target.addEventListener(event, handler)
+  return () => {
+    target.removeEventListener(event, handler)
+  }
+}