+++ webui/src/store/selectors/schema.js
@@ -1,0 +1,4 @@
+// @flow
+import type { State } from 'src/store/rootReducer';
+
+export const isValueChanged = (state: State) => state.schema.value !== state.schema.savedValue;